+++ pikoci/mysql/migrate/migrations/migrations.go
@@ -16,7 +16,7 @@ type Migration struct {
 // Migrations is the ordered list of all schema migrations. It is defined as
 // a fixed-size array so that the compiler catches ordering conflicts when
 // multiple developers add migrations concurrently.
-var Migrations = [23]Migration{
+var Migrations = [24]Migration{
 	V0Initial,
 	V1ResourceCheckInterval,
 	V2JobsAndBuilds,
@@ -40,4 +40,5 @@ var Migrations = [23]Migration{
 	V20Triggers,
 	V21PipelineCanonical,
 	V22PendingBuilds,
+	V23BackfillWebhookTokens,
 }